Friends

+++ GAMMA/src/main/java/org/game/user/controller/ConsumerBasketController.java
@@ -27,10 +27,10 @@ public class ConsumerBasketController {
 	@Autowired
 	private ConsumerBasketService service;
 	
-	@PostMapping("/basketList")
-	public String getBasketList(@RequestParam("cid") String cid, HttpSession session, Model model) {
+	@GetMapping("/basketList")
+	public String getBasketList(HttpSession session, Model model) {
 		
-		cid = (String)session.getAttribute("session_cid");
+		String cid = (String)session.getAttribute("session_cid");
 		
 		if(cid == null) {
 			return "redirect:/user/userLogin";
@@ -47,20 +47,28 @@ public class ConsumerBasketController {
 	public String additionConsumerBasket(long gnum, HttpSession session, Model model) {
 
 		String cid = (String)session.getAttribute("session_cid");
+
+		if(cid == null) {
+			return "redirect:/user/userLogin";
+		}
 		
 		service.additionConsumerBasket(cid, gnum);
 		
-		return "redirect:/gameInfo/get/" + gnum;
+		return "redirect:/gamepayment/basketList";
 	}
 	
 	@PostMapping("/basketRemove")
 	public String removeConsumerBasket(long gnum, HttpSession session, Model model) {
 		
 		String cid = (String)session.getAttribute("session_cid");
+
+		if(cid == null) {
+			return "redirect:/user/userLogin";
+		}
 		
 		service.removeConsumerBasket(cid, gnum);
 		
-		return "redirect:/gamepayment/basketList";
+		return "redirect:/gamepayment/basketList?cid=" + cid;
 	}
 	
 }
